refactor: extract ScreenCaptureKitImageProvider.swift from ScreenshotCaptureSupport.swift
Splits the `ScreenCaptureKitImageProvider` struct (~109 lines, the
concrete ScreenCaptureImageProviding implementation using SCKit) out
of ScreenshotCaptureSupport.swift into its own file. That file now
holds the small value types + the two protocols + the PNG writer
extension surface.

Byte-identical move of lines 28-136. Both files import CoreGraphics +
Foundation + @preconcurrency ScreenCaptureKit.

ScreenshotCaptureSupport.swift: 184 → 73 lines. Tests: 667.
